feat: ignore protected props in createNewSession*
